Should have renamed namespace when I renamed the class.
