+++ src/SFA.DAS.PSRService.IntegrationTests/ReportSubmission/Given_I_Have_Created_A_Report/And_I_Have_Reported_All_Mandatory_Numbers/But_Only_One_Mandatory_Factor/But_Only_One_Mandatory_Factor.cs
@@ -1,4 +1,4 @@
-﻿namespace SFA.DAS.PSRService.IntegrationTests.ReportSubmission.Given_I_Have_Created_A_Report.And_I_Have_Reported_All_Mandatory_Numbers.But_Only_Some_Mandatory_Factors
+﻿namespace SFA.DAS.PSRService.IntegrationTests.ReportSubmission.Given_I_Have_Created_A_Report.And_I_Have_Reported_All_Mandatory_Numbers.But_Only_One_Mandatory_Factor
 {
     public abstract class But_Only_One_Mandatory_Factor
     :And_I_Have_Reported_All_Mandatory_Numbers
